functions/content/process_image_variants/pyproject.toml
[tool.poetry]
name = "process_image_variants"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "2.19.0"
Pillow = "11.3.0"
functions-framework = "3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
